Split up core, add search results component, only set defaults in core
